fix(index.js): fix color modification by luminance calculation
The previous luminance function did not make a copy of the passed color array before making the
luminance calculation. As a result, the luminance function, and by extension, the contrast ratio
functions, would scale the inputted rgb values by 1/255. The new luminance creates a copy of the rgb
array before calculation, and thus avoids this problem

fix #10